remove ConfigParser
